requirements.txt
python-dotenv
langchain
openai
chromadb
tiktoken
unstructured
unstructured[pdf]
pypdf
transformers